edge case with packageURL

+++ R/pkgDep.R
@@ -307,18 +307,20 @@ pkgDepInner <- function(packages, libPath, which, keepVersionNumber,
               packageURL <- file.path(pkgName, pkgFilename)
               dt <- numeric()
             }
-            if (endsWith(packageURL, "tar.gz")) {
-              srcContrib <- "src/contrib"
-              url <- file.path(repos, srcContrib, "/Archive", packageURL) 
-              url2 <- file.path(repos, srcContrib, basename(packageURL))
-              tf <- tempfile()
-              haveFile <- suppressWarnings(tryCatch(download.file(url, tf, quiet = TRUE), error = function(x) 
-                tryCatch(download.file(url2, tf, quiet = TRUE), error = function(y) FALSE)))
-              if (file.exists(tf)) {
-                untar(tarfile = tf, exdir = td)
-                filesToDel <- dir(packageTD, recursive = TRUE, full.names = TRUE, include.dirs = TRUE)
-                filesToDel <- filesToDel[grep("^DESCRIPTION$", basename(filesToDel), invert = TRUE)]
-                unlink(filesToDel, recursive = TRUE)
+            if (!is.null(packageURL)) {
+              if (endsWith(packageURL, "tar.gz")) {
+                srcContrib <- "src/contrib"
+                url <- file.path(repos, srcContrib, "/Archive", packageURL) 
+                url2 <- file.path(repos, srcContrib, basename(packageURL))
+                tf <- tempfile()
+                haveFile <- suppressWarnings(tryCatch(download.file(url, tf, quiet = TRUE), error = function(x) 
+                  tryCatch(download.file(url2, tf, quiet = TRUE), error = function(y) FALSE)))
+                if (file.exists(tf)) {
+                  untar(tarfile = tf, exdir = td)
+                  filesToDel <- dir(packageTD, recursive = TRUE, full.names = TRUE, include.dirs = TRUE)
+                  filesToDel <- filesToDel[grep("^DESCRIPTION$", basename(filesToDel), invert = TRUE)]
+                  unlink(filesToDel, recursive = TRUE)
+                }
               }
             }
           }
